Implemented Upgrade and resource income se details
The functionality for upgrading the Town Hall should now be setup. The upgradeTownHall() function just need functionality for what it should do client side and a eventHandler to call it. 
All players now have a generateIncome function. This function just needs to be called at the start of the game, and will then keep going until the sever is shut down.